[86bzwk3eq][dropdown-menu] fixed double click to unfocus in comboboxes
